Version 3.28.59 (based on bleeding_edge revision r22888)
Removed GetConstructor from the API. Instead either get the "constructor" property stored in the prototype, or keep a side-table.

Enable ES6 Symbols by default (issue 2158).

Performance and stability improvements on all platforms.

git-svn-id: https://v8.googlecode.com/svn/trunk@22891 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/base/platform/platform-posix.cc
@@ -515,7 +515,7 @@ Thread::Thread(const Options& options)
     : data_(new PlatformData),
       stack_size_(options.stack_size()),
       start_semaphore_(NULL) {
-  if (stack_size_ > 0 && stack_size_ < PTHREAD_STACK_MIN) {
+  if (stack_size_ > 0 && static_cast<size_t>(stack_size_) < PTHREAD_STACK_MIN) {
     stack_size_ = PTHREAD_STACK_MIN;
   }
   set_name(options.name());
